Pagina Home Feita e adicionado link para cadastro

+++ CurriculifyBack/curriculify/src/main/java/com/projeto/curriculify/ExpAcademica/ExpAcademicaRepository.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ExpAcademicaRepository extends JpaRepository<ExpAcademica, Long> {
+public interface ExpAcademicaRepository extends JpaRepository<ExpAcademica, Integer> {
 	List<ExpAcademica> findAll();
 	List<ExpAcademica> findAllByIdUsuario(Integer idUsuario);
 }
